requirements.txt
# Libraries
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Machine Learning Models
xgboost==2.0.3
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web Framework for Deployment
streamlit==1.29.0

# API and Environment Management
requests==2.31.0
python-dotenv==1.0.0

# Data Processing
joblib==1.3.2

# Progress Bars
tqdm==4.66.1

# Notebook Support
jupyter==1.0.0
ipykernel==6.27.1
